DISTMYSQL-138: MySQL Orchestrator does not support new semi-sync variables introduced in MySQL 8.0.26
https://jira.percona.com/browse/DISTMYSQL-138

Problem:
MySql 8.0.26 introduced new semisync replication plugins
semisync_source.so and semisync_replica.so
in place of
semisync_master.so and semisync_slave.so
All semisync replication plugins' variables were renamed in the way
that 'master' was replaced with the term 'source' and 'slave' was
replaced by 'replica'.

As 8.0.26 contains also old plugins and when they are installed
everything works fine it is not enough to detect server version.
Moreover, it is possible to mix old source plugin with new replica
plugin and the opposite.

Solution:
Detect which plugins are installed and use proper variable names.