DISTMYSQL-208: Orchestrator GUI incorrectly shows recovery option for intermediate database in chained replication
https://jira.percona.com/browse/DISTMYSQL-208

Problem:
If we've got replication chain A->B->C, and C is down, GUI shows
'Recover' dropdown for node B, but there is no possible recovery action
available in such a case.
The problem is well visible on upstream and Percona Orchestrator up to
version v3.2.6-3. For next versions it is still there, but is hard to
reproduce.

Cause:
The cause for hard reproduction on versions > v3.2.6-6 is caused by fix
for DISTMYSQL-182. In case of the instance being unreachable, it causes
skip of instance.Key storing, which is later needed for Orchestrator's
backend update. This caused that 'last_checked' and 'last_seen'
timestamps were updated incorrectly.

The root cause of the main problem is the analysis logic in
Analysis_dao.go:GetReplicationAnalysis(). The condition for setting
AllIntermediateMasterReplicasNotReplicating does not check if there are
any replicas reachable. So the case when all replicas are dead
(no recovery action possible) and the case when some replicas are still
reachable, but are not replicating (recovery action possible) are
undistingushable.

Solution:
1. Move instance.Key assignment before Ping
2. Improve the analysis logic. Report
AllIntermediateMasterReplicasNotReplicating only if all replicas are not
replicating, but there are still some reachable replicas.