DISTMYSQL-228: Orchestrator GUI messages should optionally contain the name of the orchestratorapp generating the message
https://jira.percona.com/browse/DISTMYSQL-228

Problem:
When Orchestrator is used in HA setup it is not possible to know which
Orchestrator instance provided the response to the client. In general
it is OK however, for debugging purposes, it seems to be useful to be
able to easily identify the Orchestrator instance which serviced
the request.

Solution:
Two new configuration parameters were added
(if not present or empty strings, the old behavior applies)

PrependMessagesWithOrcIdentity:
FQDN - Fully qualified domain name of orchestrator's host will be added
with the fallback to 'hostname'
hostname - hostname of the orchestrator's host will be added with the
fallback to 'custom' if CustomOrcIdentity is not empty
custom - custom prefix as configured in CustomOrcIdentity will be added

CustomOrcIdentity:
Any string identifying Orchestrator instance. Works together with
PrependMessagesWithOrcIdentity=custom or as a fallback for
FQDN/hostname