DISTMYSQL-415: Web API errors are ignored by JS application
https://perconadev.atlassian.net/issues/DISTMYSQL-415

Problem:
When the Orchestrator's Web API returns error, UI does not react to it.
Instead of this it is in the state like the underlying action was still
in progress.

Cause:
Errors returned by Web API are not handled correctly in JavaScript
application.

Solution:
Fixed error handling in JavaScript.

Additionally:
1. Fixed nullptr exception in Web API Discover handler in case no
instances were discovered
2. Fixed error handling in ReadTopologyInstanceBufferable()