DISTMYSQL-458: Source failover promote suggestion ignored if clicked on the right side of the host
Problem:
Let's consider source with two replicas. When source is dead, UI allows
the user to trigger recovery action by selecting one of the replicas
to become the new source. The user suggestion is ignored if the user
clicks not on 'Recover, try to promote' text, but on hostname:port part.
Sometimes the selected host is selected, but sometimes it is not.

Cause:
The problem is in JavaScript application. If we look at DOM structure,
host:port part is the internal item of the whole clickable container.
Suggested host/port are passed via container attributes.
The event listener is attached to the container, so if someone clicks
on the internal item, the event bubbles up. Then we tried to get
attributes of the event target (clicked one) instead of event
currentTarget (the container), which resulted in the api call like
/api/recover/sourcehost/sourceport instead of
/api/recover/sourcehost/sourceport/suggestedhost/suggestedport

Solution:
In the event handler use the container (event listener) instead of
the clicked item as the source of needed attributes.
Additionally extended orchestrator-client. Now the 'recover' command
accepts the optional destination (suggested source successor) host.

Credits for Marcin Szalowicz for providing JS/Bash patch.