DISTMYSQL-466: RestartReplicationQuick called even from Orchestrator cluster where recovery has been globally disabled
https://perconadev.atlassian.net/browse/DISTMYSQL-466

Problem:
When recovery is disabled globally, UnreachableMasterWithLaggingReplicas
and UnreachableIntermediateMasterWithLaggingReplicas cases cause
replica io thread to be restarted.

Cause:
Commit 98bd7f0f added the feature allowing global recovery disable

Commit fc33f3ee moved the implementation to
executeCheckAndRecoverFunction

Commit b761fa3d introduced runEmergencyOperations() function.
Its purpose was to read topology instance to speed up recovery.
The instance was read, then recovery was skipped if disabled globally.

Commit 464a3c1d and 684d6e24 caused the regression. They introduced
the call to emergentlyRestartReplicationOnTopologyInstance() from
runEmergencyOperations().
https://github.com/openark/orchestrator/pull/572 and
https://github.com/openark/orchestrator/pull/1005 provide the detailed
explanation, why it was done.

Solution:
If recovery was disabled globally, and this is not forced discovery,
skip restart of replicas.

Additionally fixed Instance object read from Orchestrator's backend DB.
Such and object was missing QSP member (Query String Provider).
As the consequence any query related to master/slave <-> source/replica
could not be resolved and failed (because nil string query was executed)