DISTMYSQL-432: gtid-errant test fails when the test is done with PS 8.0.x
https://perconadev.atlassian.net/issues/DISTMYSQL-432

Cause:
The test fails, because the tested feature is not supported by 8.0
server version.

Solution:
Added optional 'run_condition' script that can be located in the test's
main directory. Its purpose is to do any validation needed and if it
exits with nonzero code, the test is skipped by the test runner.