DISTMYSQL-524: Orchestrator does not downtime the instance without the time provided in API
https://perconadev.atlassian.net/browse/DISTMYSQL-524

Problem:
api/begin-downtime endpoint allows setting the node in downtime mode
without specifying the downtime time. In such a case the node is not
set in downtime mode.

Cause:
If downtime time is not specified, a downtime is set internally to zero.
Then setting it is skipped, because the flow decides that it is expired
anyway.

Solution:
In case of no downtime time specified, fall back to the default, hard
codded value of MaintenanceExpireMinutes (10 minutes)

Thanks for the investigation and th original contribution
by Yunus Shaikh.